+++ custom_components/electricity_pro/const.py
@@ -48,6 +48,7 @@ CONF_FORECAST_NORDPOOL_CONFIG_ENTRY = "forecast_nordpool_config_entry"
 CONF_PRICING_STRATEGY = "pricing_strategy"
 CONF_PRICE_INCLUDED_COMPONENTS = "price_included_components"
 CONF_PRICE_VAT_TREATMENT = "price_vat_treatment"
+CONF_PRICE_VAT_RATE = "price_vat_rate"
 CONF_PRICE_COMPLETENESS = "price_completeness"
 
 PLATFORMS: list[Platform] = [Platform.SENSOR, Platform.BINARY_SENSOR]